Refactor imports and clean up unused files
- Removed the `badActors.ts` and `interfaces.i.ts` files to streamline the codebase and eliminate redundancy.
- Updated import statements across various modules to ensure consistency and compatibility with the new structure, particularly in `anticheat_logs.ts`, `ServiceManager.ts`, and `message_handler.ts`.
- Enhanced the organization of type imports from `interfaces.ts`, improving clarity and maintainability.

These changes improve the overall structure of the codebase, facilitating easier navigation and future enhancements.
