--- src/badActors.ts
@@ -1,1 +1,0 @@
-export const idList: string[] = ["804933185216577566"];
