+++ src/player_device_listener/playerDeviceLogging.ts
@@ -1,7 +1,6 @@
-import { MessagePayload, MessageCreateOptions } from "discord.js";
-import { EmbedBuilder, TextBasedChannel } from "discord.js";
-import { loadConfig } from "../core/config/configLoader.js";
 import { Client } from "bedrock-protocol";
+import { EmbedBuilder, MessageCreateOptions, MessagePayload, TextBasedChannel } from "discord.js";
+import { loadConfig } from "../core/config/configLoader.js";
 
 /* Add to prevent the message being spammed this will allow the blacklist to work, 
 it seems that when a player leaves the range of the bot account
