+++ src/utils/discord_client.ts
@@ -1,11 +1,11 @@
 import { Client, GatewayIntentBits, TextBasedChannel } from "discord.js";
 
-import { registerCommands } from "../translation/registerCommands.js";
-import { handleCommand, handleChatMessage } from "./message_handler.js";
-import { handleWhitelistCommand } from "./whitelist_manager.js";
-import { checkAndDeleteEmptyChannels } from "../voiceChat_listener/voiceChatCleanUp.js";
 import { ConfigTemplate } from "../core/config/types.js";
-import { IChannelConfig } from "../interface/interfaces.i.js";
+import { IChannelConfig } from "../core/types/interfaces.js";
+import { registerCommands } from "../translation/registerCommands.js";
+import { checkAndDeleteEmptyChannels } from "../voiceChat_listener/voiceChatCleanUp.js";
+import { handleChatMessage, handleCommand } from "./message_handler.js";
+import { handleWhitelistCommand } from "./whitelist_manager.js";
 
 /**
  * Creates and sets up the Discord client with proper intents
