--- src/utils/logger.ts
@@ -1,8 +1,0 @@
-export const logger = {
-    error: (message: string, error?: any) => {
-        console.error(`[ERROR] ${message}`, error || "");
-    },
-    info: (message: string) => {
-        console.log(`[INFO] ${message}`);
-    },
-};
