+++ src/utils/text_corrections.ts
@@ -1,8 +1,8 @@
 import { readFileSync } from "fs";
 import { join } from "path";
-import { CSZETranslations } from "./CSZETranslations.js";
-import { ICorrectionMap, ILocalizationFile } from "../interface/interfaces.i.js";
 import { logger } from "../core/logging/logger.js";
+import { ICorrectionMap, ILocalizationFile } from "../core/types/interfaces.js";
+import { CSZETranslations } from "./CSZETranslations.js";
 
 /**
  * Translates text from TTX encoding back to regular text
